oxygenConcentrationIdx: make unsigned

+++ opm/models/blackoil/blackoilnewtonmethod.hpp
@@ -416,7 +416,7 @@ protected:
                                                   this->problem().referencePorosity(globalDofIdx, 0) - 1e-8);
                 }
                 if constexpr (enableMICP) {
-                    if (pvIdx == Indices::oxygenConcentrationIdx) {
+                    if (pvIdx == static_cast<int>(Indices::oxygenConcentrationIdx)) {
                         nextValue[pvIdx] = std::max(nextValue[pvIdx], Scalar{0.0});
                     }
                     if (pvIdx == Indices::ureaConcentrationIdx) {
